# apps/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/vehicles/{vehicleId}/cost-summary": {
      "get": {
        "operationId": "CostSummaryController_getSummary_v1",
        "parameters": [
          {
            "name": "vehicleId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "First month, inclusive. Default: 11 months before `to`.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "description": "Last month, inclusive. Default: the current UTC month.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CostSummaryResponseDto"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "cost-summary"
        ]
      }
    },
    "/api/v1/cost-summary": {
      "get": {
        "operationId": "FleetCostSummaryController_getFleetSummary_v1",
        "parameters": [
          {
            "name": "from",
            "required": false,
            "in": "query",
            "description": "First month, inclusive. Default: 11 months before `to`.",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "description": "Last month, inclusive. Default: the current UTC month.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CostSummaryResponseDto"
                }
              }
            }
          },
          "default": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "cost-summary"
        ]
      }
    },
...
  },
  "info": {
    "title": "FleetOps API",
    "description": "",
    "version": "v1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ValidationErrorDetailDto": {
        "type": "object",
        "properties": {
          "field": {
            "type": "string",
            "description": "Dotted path of the invalid field, for example `items.0.name`."
          },
          "messages": {
            "description": "Validation messages for the field.",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "field",
          "messages"
        ]
      },
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "description": "HTTP status code."
          },
          "error": {
            "type": "string",
            "description": "HTTP reason phrase, for example `Not Found`."
          },
          "message": {
            "type": "string",
            "description": "Human readable summary."
          },
          "requestId": {
            "type": "string",
            "description": "Equals the X-Request-Id response header."
          },
          "timestamp": {
            "type": "string",
            "description": "ISO 8601 time the error was produced."
          },
          "path": {
            "type": "string",
            "description": "Request path without the query string."
          },
          "details": {
            "description": "Per-field problems. Only present on validation errors.",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ValidationErrorDetailDto"
            }
          }
        },
        "required": [
          "statusCode",
          "error",
          "message",
          "requestId",
          "timestamp",
          "path"
        ]
      },
...
      "CostSummaryMonthDto": {
        "type": "object",
        "properties": {
          "month": {
            "type": "string",
            "description": "\"YYYY-MM\""
          },
          "maintenanceCost": {
            "type": "string"
          },
          "fuelCost": {
            "type": "string"
          },
          "fuelLiters": {
            "type": "string"
          },
          "totalCost": {
            "type": "string"
          }
        },
        "required": [
          "month",
          "maintenanceCost",
          "fuelCost",
          "fuelLiters",
          "totalCost"
        ]
      },
      "CostSummaryTotalsDto": {
        "type": "object",
        "properties": {
          "maintenanceCost": {
            "type": "string"
          },
          "fuelCost": {
            "type": "string"
          },
          "fuelLiters": {
            "type": "string"
          },
          "totalCost": {
            "type": "string"
          }
        },
        "required": [
          "maintenanceCost",
          "fuelCost",
          "fuelLiters",
          "totalCost"
        ]
      },
      "CostSummaryResponseDto": {
        "type": "object",
        "properties": {
          "from": {
            "type": "string"
          },
          "to": {
            "type": "string"
          },
          "months": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CostSummaryMonthDto"
            }
          },
          "totals": {
            "$ref": "#/components/schemas/CostSummaryTotalsDto"
          }
        },
        "required": [
          "from",
          "to",
          "months",
          "totals"
        ]
      },
...
    }
  }
}